Khalansky: Revert making % unsigned
So, after debugging, finding out which behavior of my system made
it fail the extra tests, and fixing the problems, I try to push my
results, only to find out that my push doesn't go through: someone
else committed just now.

Who would've thought: it was the author of the tests, removing one
of the tests that made life difficult. So, reverting one of the
most controversial changes.